fixed nasty nextId of event bug

+++ app/jvm/src/main/scala/pl/setblack/lsa/io/FileStorage.scala
@@ -1,17 +1,17 @@
 package pl.setblack.lsa.io
 
 import java.io.ObjectOutputStream
-
-import scala.collection.JavaConverters._
 import java.nio.file.{Files, Paths, StandardOpenOption}
 import java.util.regex.Pattern
 
 import pl.setblack.lsa.io.DataStorage.{DataInputStream, DataOutputStream, DataStorage}
 
+import scala.collection.JavaConverters._
 import scala.concurrent.{ExecutionContext, Future}
 
 class FileStorage(val diskPath: String)(implicit val ectx: ExecutionContext) extends DataStorage {
   val filePattern = Pattern.compile("events_\\d+")
+
   override def openDataReader(path: Seq[String]): Future[Option[DataInputStream]] = {
     Future {
       val fsPath = createPath(path)
@@ -19,12 +19,12 @@ class FileStorage(val diskPath: String)(implicit val ectx: ExecutionContext) ext
         val directoryStream = Files.newDirectoryStream(fsPath)
         val filesInDirectory = directoryStream.iterator().asScala.toList
 
-        val sorted =   ( filesInDirectory
-          filter( file => filePattern.matcher(file.getFileName.toString).matches())
-          map(file => (file.getFileName.toString.substring("events_".length),file))
-          map{ case (numer, file) => (numer.toInt, file) }
-          sortBy{ case (numer, file) => numer}
-          map{ case (numer, file) => file }
+        val sorted = (filesInDirectory
+          filter (file => filePattern.matcher(file.getFileName.toString).matches())
+          map (file => (file.getFileName.toString.substring("events_".length), file))
+          map { case (numer, file) => (numer.toInt, file) }
+          sortBy { case (numer, file) => numer }
+          map { case (numer, file) => file }
           )
 
         Some(new FileInputStream(sorted.iterator))
@@ -36,21 +36,22 @@ class FileStorage(val diskPath: String)(implicit val ectx: ExecutionContext) ext
 
   override def openDataWriter(path: Seq[String]): Future[DataOutputStream] = {
     Future {
+      println(s"opening writer to ${path}")
       val fsPath = createPath(path)
       if (!Files.exists(fsPath)) {
         Files.createDirectories(fsPath)
       }
       val directoryStream = Files.newDirectoryStream(fsPath)
       val filesInDirectory = directoryStream.iterator().asScala.toList
-      val maxIndex = ( ( filesInDirectory
-          filter( file => filePattern.matcher(file.getFileName.toString).matches())
-          map(file => file.getFileName.toString.substring("events_".length))
-          map(_.toInt) ) :+ 0
-          max
+      val maxIndex = ((filesInDirectory
+        filter (file => filePattern.matcher(file.getFileName.toString).matches())
+        map (file => file.getFileName.toString.substring("events_".length))
+        map (_.toInt)) :+ 0
+        max
         )
-       val newFile = fsPath.resolve("events_" +(maxIndex + 1))
-       val outputStream = Files.newOutputStream(newFile, StandardOpenOption.CREATE_NEW)
-       val objectOut  = new ObjectOutputStream(outputStream)
+      val newFile = fsPath.resolve("events_" + (maxIndex + 1))
+      val outputStream = Files.newOutputStream(newFile, StandardOpenOption.CREATE_NEW)
+      val objectOut = new ObjectOutputStream(outputStream)
       new FileOutputStream(objectOut)
     }
   }
@@ -62,8 +63,11 @@ class FileStorage(val diskPath: String)(implicit val ectx: ExecutionContext) ext
 
   def erase(path: Seq[String]) = {
     val fsPath = createPath(path)
-    val directoryStream = Files.newDirectoryStream(fsPath)
-    val filesInDirectory = directoryStream.iterator().asScala.toList
-    filesInDirectory.foreach( f=> Files.delete(f))
+    if (Files.exists(fsPath)) {
+      val directoryStream = Files.newDirectoryStream(fsPath)
+      val filesInDirectory = directoryStream.iterator().asScala.toList
+      filesInDirectory.foreach(f => Files.delete(f))
+    }
   }
+
 }
